Scheduler: Major overhaul of Scheduler configuration to make it compatible with the latest version of the Scheduler.
The scheduler will now enforce that block names follow the BLOCK-000 (for block tickets) or BLOCK-T000 (for block test cases) name format.
This requires changing several block names to match this new convention.

Also, this commit removes all the blocks from OR4 and OR3 since we are now unlikely to run them again.
They were also not following the naming convention and it would not be worth the effort to fix them.